pyproject.toml
[project]
name = "wialonblock"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "aiogram>=3.20.0.post0",
    "py-aiowialon>=1.3.4.post1",
]

[tool.uv]
package = true

[project.scripts]
wialonblock = "wialonblock.__main__:main"